Changed the way the compiler operates to be able to handle array locations

+++ src/edu/mit/compilers/cfg/CFGAble.java
@@ -15,5 +15,5 @@ public interface CFGAble {
      *  	onto the stack for future expressions that want to use it's value
      */
     public void generateAssembly(AssemblyContext ctx);
-    public int getNumStackAllocations();
+    public long getNumStackAllocations();
 }
