+++ src/edu/mit/compilers/cfg/components/LeaveBlock.java
@@ -14,7 +14,7 @@ import edu.mit.compilers.lowir.instructions.Jmp;
 import edu.mit.compilers.lowir.instructions.Label;
 
 public class LeaveBlock extends BasicBlock {
-	private int numStackAllocations;
+	private long numStackAllocations;
 
     public LeaveBlock(List<CFGAble> components) {
         super(components);
@@ -28,8 +28,8 @@ public class LeaveBlock extends BasicBlock {
     public boolean isEmpty() {
         return false;
     }
-    
-    public void setNumStackAllocations(int numStackAllocations) {
+
+    public void setNumStackAllocations(long numStackAllocations) {
     	this.numStackAllocations = numStackAllocations;
     }
